pyproject.toml
[tool.poetry]
name = "scrape_em_all"
version = "0.1.0"
description = ""
authors = ["Valentyn Slivko <metallistvalon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.21"
celery = "^5.2.7"
redis = "^4.3.4"
python-dotenv = "^0.20.0"
aiohttp = "^3.8.1"
beautifulsoup4 = "^4.11.1"
mongoengine = "^0.24.2"
python-dateutil = "^2.8.2"
flower = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
